requirements.txt
# python==3.11.5
langchain==0.0.283
openai
docarray==0.37.1
pypdf==3.15.5
pinecone-client==2.2.2
python-dotenv==1.0.0
guidance==0.1.3
tiktoken==0.5.1